Add information about generating minified versions
It should be possible to audit the patches and create reproducible
builds. This can be only achieved by applying the patches for the
unminified versions and then minify them using the uglify-js CLI which
requires the correct version and flags for reproducible builds.

To make this easier I have created minify.sh which contains the command
for all versions.